# docs/swagger.yaml
basePath: /v1
definitions:
...
  domainuser.UserRole:
    enum:
    - admin
    - cashier
    type: string
    x-enum-varnames:
    - Admin
    - Cashier
...
  modelv1.ErrorResponse:
    properties:
      messages:
        example:
        - Error message 1
        - ' Error message 2'
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
...
  modelv1.Meta:
    properties:
      limit:
        example: 10
        type: integer
      skip:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  modelv1.RegisterRequest:
    properties:
      email:
        example: test@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: "12345678"
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  modelv1.Response:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  modelv1.UpdateUserRequest:
    properties:
      email:
        example: test@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: "12345678"
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/domainuser.UserRole'
        example: admin
    required:
    - email
    - name
    - password
    - role
    type: object
  modelv1.UserResponse:
    properties:
      created_at:
        example: "1970-01-01T00:00:00Z"
        type: string
      email:
        example: test@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      updated_at:
        example: "1970-01-01T00:00:00Z"
        type: string
    type: object
host: localhost
info:
  contact:
    email: letienminh2512@gmail.com
    name: Tien Minh
    url: https://github.com/TienMinh25/go-hexagonal-architecture
  description: This is a simple RESTful Point of Sale (POS) Service API written in
    Go using Gin web framework, PostgreSQL database, and Redis cache.
  license:
    name: APACHE
    url: https://github.com/TienMinh25/go-hexagonal-architecture/blob/main/LICENSE
  title: Go Sale API demo for hexagonal architecture
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: List users with pagination
      parameters:
      - description: Skip
        in: query
        name: skip
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users displayed
          schema:
            $ref: '#/definitions/modelv1.Meta'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: create a new user account with default role "cashier"
      parameters:
      - description: Register request
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/modelv1.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User created
          schema:
            $ref: '#/definitions/modelv1.UserResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "409":
          description: Data conflict error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      summary: Register a new user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            $ref: '#/definitions/modelv1.Response'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get a user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User displayed
          schema:
            $ref: '#/definitions/modelv1.UserResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update a user's name, email, password, or role by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update user request
        in: body
        name: updateUserRequest
        required: true
        schema:
          $ref: '#/definitions/modelv1.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            $ref: '#/definitions/modelv1.UserResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "403":
          description: Forbidden error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "404":
          description: Data not found error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/modelv1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - Users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"